requirements.txt
fitz==0.0.1.dev2
graphviz==0.20.3
langchain==0.3.20
langchain_chroma==0.2.2
langchain_community==0.3.19
langchain_core==0.3.41
langchain_groq==0.2.4
langchain_huggingface==0.1.2
langchain_openai==0.3.7
langchain_text_splitters==0.3.6
matplotlib==3.10.1
openai==1.65.4
pandas==2.2.3
Pillow==11.1.0
python-dotenv==1.0.1
streamlit==1.40.2
streamlit_drawable_canvas==0.9.3
streamlit_image_coordinates==0.1.9
frontend
pypdf
